merchandising-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.1
python-multipart==0.0.6
azure-search-documents==11.4.0
